# services/order-accept/src/OrderAccept.Application/Contracts/Requests/CreateOrderRequest.cs
namespace OrderAccept.Application.Contracts.Requests;

public sealed record CreateOrderRequest(
    IReadOnlyCollection<CreateOrderItem> Items
);

public sealed record CreateOrderItem(
    string ProductId,
    int Quantity
);